add to db in diaglog fixes

+++ src/db/controllers/sqlite3.controller.ts
@@ -5,31 +5,48 @@ import * as Entities from "../entities";
 export class Sqlite3Controller {
 	private repositories = {};
     constructor() {
-		Object.keys(Entities).forEach(name => {
-			this.repositories[name.toLowerCase()] = getRepository(Entities[name]);
-		});
+		this.saveEntityRepos();
 	}
-    private getRepoFromRequest(request:Request) {
-        let entityName = request.route.path.replace("/","").toLowerCase();
-        return this.repositories[entityName];
-    }
 	async all(request: Request, response: Response, next: NextFunction) {
 		let repo = this.getRepoFromRequest(request);
-		return repo.find();
+		if (repo) return repo.find();
     }
     async one(request: Request, response: Response, next: NextFunction) {
 		let repo = this.getRepoFromRequest(request);
-        return repo.findOne(request.params.id);
+		if (repo) return repo.findOne(request.params.id);
+
     }
     async save(request: Request, response: Response, next: NextFunction) {
         let repo = this.getRepoFromRequest(request);
-		await repo.save(request.body);
-		return repo.find();
+		if (repo) {
+			await repo.save(request.body);
+			return repo.find();
+		}
     }
     async remove(request: Request, response: Response, next: NextFunction) {
 		let repo = this.getRepoFromRequest(request);
-        const itemToRemove = await repo.findOne(request.params.id);
-        await repo.remove(itemToRemove);
-		return repo.find();
+		if (repo) {
+			const itemToRemove = await repo.findOne(request.params.id);
+			await repo.remove(itemToRemove);
+			return repo.find();
+		}
+    }
+	private saveEntityRepos() {
+		Object.keys(Entities).forEach(name => {
+			if (Entities[name]) {
+				this.repositories[name.toLowerCase()] = getRepository(Entities[name])
+			}
+		});
+	}
+    private getRepoFromRequest(request:Request) {
+		let repo;
+		let preceedingSlash = /(^\/)/;
+		let entityName = request.route.path.replace(preceedingSlash, "").split("/")[0];
+		try {
+			repo = this.repositories[entityName];
+		} catch(e) {
+			console.error(e)
+		}
+		return repo;
     }
 }
